Separated FaultManager and Fault classes
