+++ FSW/src/util/fault.cpp
@@ -1,0 +1,22 @@
+/* fault.cpp defines the Fault class
+ * Usage:
+ *  module functionality
+ *  function definitions
+ *  put function declarations in memUtil.hpp
+ * 
+ * Modules encompassed:
+ *  Fault Detection
+ *  Fault Logging
+ *  Fault Mitigation
+ *
+ * Additional files needed for compilation:
+ *  config.hpp
+ *  
+ */
+
+/* - - - - - - Includes - - - - - - */
+// All libraries are put in dataCollection.hpp
+// NS2 headers
+#include "../headers/fault.hpp"
+
+
